move extra settings
