+++ management/server/integrations/extra_settings/manager.go
@@ -3,10 +3,10 @@ package extra_settings
 import (
 	"context"
 
-	"github.com/netbirdio/netbird/management/server/account"
+	"github.com/netbirdio/netbird/management/server/types"
 )
 
 type Manager interface {
-	GetExtraSettings(ctx context.Context, accountID string) (*account.ExtraSettings, error)
-	UpdateExtraSettings(ctx context.Context, accountID string, extraSettings *account.ExtraSettings) error
+	GetExtraSettings(ctx context.Context, accountID string) (*types.ExtraSettings, error)
+	UpdateExtraSettings(ctx context.Context, accountID string, extraSettings *types.ExtraSettings) error
 }
